Clip2/backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1

# OpenAI integration
openai==0.28

# Text analysis for sentiment
textblob==0.17.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0

# CORS for frontend communication
fastapi-cors==0.0.6 